more simplifications to bokeh module
issues: closes #2051
