+++ bokeh/browserlib.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from os.path import abspath
 import webbrowser
 
-from . import settings
+from .settings import settings
 
 def get_browser_controller(browser=None):
     browser = settings.browser(browser)
